requirements.txt
bitsandbytes==0.35.4
decord==0.6.0
nerfacc==0.3.3
pyhocon==0.3.57
icecream==2.1.0
PyMCubes==0.1.2
modelcards
einops
ftfy
piq
imageio
imageio-ffmpeg
pyransac3d
torch_efficient_distloss
rembg
pytorch-lightning==1.9.5